Connected Spring application with Docker Postgres
So now it is also fixed that when run spring application it connects to Postgres in Docker.
Just remember that if you do "docker run -p 8081:4000 --name reservation-system-v1 reservation-system:v1" in command it will still cause issue.
BUT if you execute "docker-compose -f docker-compose.yml up" it will work smoothly, such that spring app will connect to Postgres inside Docker. Also, application properties is also updated to facililate. Actually application.properties is the most important part in this whole process. It is like a glue which is configuration file to package the spring app and expose the port.